Added CPU instruction RTI.
Also:
- Fixed timing errors in PLA and PLP instructions.
- Fixed pop16() method of the CPU since it previously returned 8 bits.

+++ include/nesturbia/cpu.hpp
@@ -61,7 +61,7 @@ struct Cpu {
   void write16(uint16 address, uint16 value);
 
   uint8 pop();
-  uint8 pop16();
+  uint16 pop16();
 
   void push(uint8 value);
   void push16(uint16 value);
